+++ cs-20151215/alibabacloud_cs20151215/models/_describe_cluster_node_pools_response_body.py
@@ -1572,9 +1572,11 @@ class DescribeClusterNodePoolsResponseBodyNodepoolsManagementAutoRepairPolicy(Da
     def __init__(
         self,
         approval_required: bool = None,
+        auto_repair_policy_id: str = None,
         restart_node: bool = None,
     ):
         self.approval_required = approval_required
+        self.auto_repair_policy_id = auto_repair_policy_id
         # Specifies whether to allow node restart. This parameter takes effect only when `auto_repair=true` is specified.
         # 
         # *   `true`: allows node restart.
@@ -1592,6 +1594,9 @@ class DescribeClusterNodePoolsResponseBodyNodepoolsManagementAutoRepairPolicy(Da
         if self.approval_required is not None:
             result['approval_required'] = self.approval_required
 
+        if self.auto_repair_policy_id is not None:
+            result['auto_repair_policy_id'] = self.auto_repair_policy_id
+
         if self.restart_node is not None:
             result['restart_node'] = self.restart_node
 
@@ -1601,6 +1606,9 @@ class DescribeClusterNodePoolsResponseBodyNodepoolsManagementAutoRepairPolicy(Da
         m = m or dict()
         if m.get('approval_required') is not None:
             self.approval_required = m.get('approval_required')
+
+        if m.get('auto_repair_policy_id') is not None:
+            self.auto_repair_policy_id = m.get('auto_repair_policy_id')
 
         if m.get('restart_node') is not None:
             self.restart_node = m.get('restart_node')