+++ cs-20151215/alibabacloud_cs20151215/models/_modify_cluster_node_pool_request.py
@@ -1159,9 +1159,11 @@ class ModifyClusterNodePoolRequestManagementAutoRepairPolicy(DaraModel):
     def __init__(
         self,
         approval_required: bool = None,
+        auto_repair_policy_id: str = None,
         restart_node: bool = None,
     ):
         self.approval_required = approval_required
+        self.auto_repair_policy_id = auto_repair_policy_id
         # Specifies whether ACK is allowed to automatically restart nodes after repairing the nodes. Valid values:
         # 
         # *   `true`: yes.
@@ -1179,6 +1181,9 @@ class ModifyClusterNodePoolRequestManagementAutoRepairPolicy(DaraModel):
         if self.approval_required is not None:
             result['approval_required'] = self.approval_required
 
+        if self.auto_repair_policy_id is not None:
+            result['auto_repair_policy_id'] = self.auto_repair_policy_id
+
         if self.restart_node is not None:
             result['restart_node'] = self.restart_node
 
@@ -1188,6 +1193,9 @@ class ModifyClusterNodePoolRequestManagementAutoRepairPolicy(DaraModel):
         m = m or dict()
         if m.get('approval_required') is not None:
             self.approval_required = m.get('approval_required')
+
+        if m.get('auto_repair_policy_id') is not None:
+            self.auto_repair_policy_id = m.get('auto_repair_policy_id')
 
         if m.get('restart_node') is not None:
             self.restart_node = m.get('restart_node')